5 commit

+++ app/src/main/java/com/example/todo/view/AddFragment.kt
@@ -35,7 +35,7 @@ class AddFragment : Fragment() {
         addButton.setOnClickListener {
             val title = titleEditText.text.toString()
             val des = desEditText.text.toString()
-            val date = dateEditText.text.toString().toInt()
+            val date = dateEditText.text.toString()
            // val dateButton = dateButton
 
             // how i put the date button ?? or i dont need it
